requirements.txt
openai>=1.3.0
python-dotenv>=1.0.0
google-cloud-texttospeech>=2.14.1
google-cloud-storage>=2.10.0
gTTS>=2.5.0
requests>=2.31.0

